Datauri support (#51)
* Creating DataURI for image inlining

Should solve the issue for this preprocesssor and create a better
user-experience. Also enables other downstream renderers to work
agnostically and not only the html renderer. Can convert text files to
data-uri but will still defer to the original text embedding method to
preserve historical functionality

- https://github.com/rust-lang/mdBook/issues/1087

* Made the use of data URIs configurable

* Cargo fmt and compiler warnings

* Do not create data uri images in book src dir
Create them in the .mdbook-plantuml-cache dir in the book root instead

* Update e2e tests

* Use markdown output instead of <img> tags and allow clicking
Clicking appears to be blocked for data uris on most browsers

Co-authored-by: Paul FREAKN Baker <paul-nelson-baker@users.noreply.github.com>